Convert to multiloader project, using jaredlll08's template
