+++ .gitignore
@@ -1,36 +1,27 @@
 # custom sodium
 custom_sodium/
 
-# gradle
-
-.gradle/
-build/
-out/
-classes/
-
 # eclipse
-
+bin
 *.launch
-
-# idea
-
-.idea/
-*.iml
-*.ipr
-*.iws
-
-# vscode
-
-.settings/
-.vscode/
-bin/
+.settings
+.metadata
 .classpath
 .project
 
-# macos
+# idea
+out
+*.ipr
+*.iws
+*.iml
+.idea/*
+!.idea/scopes
 
-*.DS_Store
+# gradle
+build
+.gradle
 
-# fabric
-
-run/
+# other
+eclipse
+run
+runs
