+++ common/src/main/java/rynnavinx/sspb/common/services/IPlatformHelper.java
@@ -1,0 +1,11 @@
+package rynnavinx.sspb.common.services;
+
+import java.nio.file.Path;
+
+
+public interface IPlatformHelper {
+
+    IPlatformHelper INSTANCE = Services.load(IPlatformHelper.class);
+
+    Path getConfigDirectory();
+}
