+++ common/src/main/java/rynnavinx/sspb/common/services/Services.java
@@ -1,0 +1,23 @@
+package rynnavinx.sspb.common.services;
+
+import rynnavinx.sspb.common.client.SSPBClientMod;
+
+import java.util.ServiceLoader;
+
+
+// Service loaders are a built-in Java feature that allow us to locate implementations of an interface that vary from one
+// environment to another. In the context of MultiLoader we use this feature to access a mock API in the common code that
+// is swapped out for the platform specific implementation at runtime.
+public class Services {
+
+    // This code is used to load a service for the current environment. Your implementation of the service must be defined
+    // manually by including a text file in META-INF/services named with the fully qualified class name of the service.
+    // Inside the file you should write the fully qualified class name of the implementation to load for the platform.
+    public static <T> T load(Class<T> clazz) {
+        final T loadedService = ServiceLoader.load(clazz)
+                .findFirst()
+                .orElseThrow(() -> new NullPointerException("Failed to load service for " + clazz.getName()));
+        SSPBClientMod.LOGGER.debug("Loaded {} for service {}", loadedService, clazz);
+        return loadedService;
+    }
+}
