+++ common/src/main/resources/META-INF/accesstransformer.cfg
@@ -1,0 +1,1 @@
+public net.minecraft.client.renderer.block.ModelBlockRenderer$AmbientOcclusionFace
