+++ common/src/main/resources/assets/sspb/icon.png
