+++ common/src/main/resources/sspb.accesswidener
