+++ common/src/main/resources/sspb.mixins.json
@@ -1,0 +1,18 @@
+{
+  "required": true,
+  "minVersion": "0.8",
+  "package": "rynnavinx.sspb.common.mixin",
+  "refmap": "${mod_id}.refmap.json",
+  "compatibilityLevel": "JAVA_17",
+  "client": [
+    "minecraft.AmbientOcclusionFaceAccessor",
+    "minecraft.ModelBlockRendererAccessor",
+    "minecraft.frapi.MixinModelBlockRenderer",
+    "sodium.AoFaceDataAccessor",
+    "sodium.MixinSmoothLightPipeline",
+    "sodium.MixinSodiumOptionsGUI"
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
