+++ fabric/src/main/resources/META-INF/services/rynnavinx.sspb.common.services.IPlatformHelper
@@ -1,0 +1,1 @@
+rynnavinx.sspb.fabric.services.FabricPlatformHelper
