+++ fabric/src/main/resources/fabric.mod.json
@@ -1,0 +1,45 @@
+{
+  "schemaVersion": 1,
+  "id": "${mod_id}",
+  "version": "${version}",
+
+  "name": "${mod_name}",
+  "description": "${description}",
+  "authors": [
+    "${mod_author}"
+  ],
+  "contact": {
+    "sources": "https://github.com/Rynnavinx/sodium-shadowy-path-blocks",
+    "issues": "https://github.com/Rynnavinx/sodium-shadowy-path-blocks/issues"
+  },
+
+  "license": "${license}",
+  "icon": "assets/sspb/icon.png",
+
+  "custom": {
+    "modmenu": {
+      "links": {
+        "modmenu.modrinth": "https://modrinth.com/mod/sodium-shadowy-path-blocks",
+        "modmenu.curseforge" : "https://www.curseforge.com/minecraft/mc-mods/sodium-shadowy-path-blocks"
+      }
+    }
+  },
+
+  "environment": "client",
+  "entrypoints": {
+    "client": [
+      "rynnavinx.sspb.fabric.client.SSPBFabricClientMod"
+    ]
+  },
+  "mixins": [
+    "sspb.mixins.json"
+  ],
+
+  "accessWidener" : "sspb.accesswidener",
+
+  "depends": {
+    "fabricloader": ">=0.12.10",
+    "java": ">=${java_version}",
+    "sodium": ">=0.6.0-"
+  }
+}
