+++ gradle.properties
@@ -1,20 +1,44 @@
-# Done to increase the memory available to gradle.
-org.gradle.jvmargs=-Xmx1G
+# Important Notes:
+# Every field you add must be added to the root build.gradle expandProps map.
 
-# Fabric Properties
-    # check these on https://fabricmc.net/versions.html
-    minecraft_version=1.21
-    loader_version=0.15.11
+# Project
+version=4.0.0-multiloader_test
+group=rynnavinx
+archives_base_name=sodium-shadowy-path-blocks
+java_version=17
 
-# Parchment Properties
-    parchment_minecraft=1.21
-    parchment_version=2024.06.23
+# Common
+minecraft_version=1.21
+mod_name=Sodium Shadowy Path Blocks
+mod_author=Rynnavinx
+mod_id=sspb
+license=LGPLv3
+credits=
+description=SSPB reintroduces vanilla-like smooth lighting to non-full blocks (e.g. dirt paths and how they have dark shading when directly next to a full block) when using Sodium.
+minecraft_version_range=
+## This is the version of minecraft that the 'common' project uses, you can find a list of all versions here
+## https://projects.neoforged.net/neoforged/neoform
+neo_form_version=1.20.4-20240627.114801
+# The version of ParchmentMC that is used, see https://parchmentmc.org/docs/getting-started#choose-a-version for new versions
+parchment_minecraft=1.21
+parchment_version=2024.06.23
 
-# Mod Properties
-    mod_version = 3.4.0
-    maven_group = rynnavinx
-    archives_base_name = sodium-shadowy-path-blocks
+# Fabric
+fabric_version=0.101.0+1.21
+fabric_loader_version=0.15.11
 
-# Dependencies
-    sodium_version=0.6.0-alpha.3+mc1.21
-    fabric_api_version=0.101.0+1.21
+# Forge
+forge_version=51.0.17
+forge_loader_version_range=[51,)
+
+# NeoForge
+neoforge_version=21.0.37-beta
+neoforge_loader_version_range=[4,)
+
+# Sodium
+sodium_version=0.6.0-alpha.3+mc1.21
+
+# Gradle
+org.gradle.jvmargs=-Xmx3G
+org.gradle.daemon=false
+org.gradle.parallel=true
