--- src/main/resources/fabric.mod.json
@@ -1,45 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "sspb",
-  "version": "${version}",
-
-  "name": "Sodium Shadowy Path Blocks",
-  "description": "SSPB reintroduces vanilla-like smooth lighting to non-full blocks (e.g. dirt paths and how they have dark shading when directly next to a full block) when using Sodium.",
-  "authors": [
-    "Rynnavinx"
-  ],
-  "contact": {
-    "sources": "https://github.com/Rynnavinx/sodium-shadowy-path-blocks",
-    "issues": "https://github.com/Rynnavinx/sodium-shadowy-path-blocks/issues"
-  },
-
-  "license": "LGPLv3",
-  "icon": "assets/sspb/icon.png",
-
-  "custom": {
-    "modmenu": {
-      "links": {
-        "modmenu.modrinth": "https://modrinth.com/mod/sodium-shadowy-path-blocks",
-        "modmenu.curseforge" : "https://www.curseforge.com/minecraft/mc-mods/sodium-shadowy-path-blocks"
-      }
-    }
-  },
-
-  "environment": "client",
-  "entrypoints": {
-    "client": [
-      "rynnavinx.sspb.client.SSPBClientMod"
-    ]
-  },
-  "mixins": [
-    "sspb.mixins.json"
-  ],
-
-  "accessWidener" : "sspb.accesswidener",
-
-  "depends": {
-    "fabricloader": ">=0.12.10",
-    "java": ">=17",
-    "sodium": ">=0.6.0-"
-  }
-}
