--- src/main/resources/sspb.mixins.json
@@ -1,17 +1,0 @@
-{
-  "required": true,
-  "minVersion": "0.8",
-  "package": "rynnavinx.sspb.mixin",
-  "compatibilityLevel": "JAVA_17",
-  "client": [
-    "minecraft.AmbientOcclusionFaceAccessor",
-    "minecraft.ModelBlockRendererAccessor",
-    "minecraft.frapi.MixinModelBlockRenderer",
-    "sodium.AoFaceDataAccessor",
-    "sodium.MixinSmoothLightPipeline",
-    "sodium.MixinSodiumOptionsGUI"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
